Add files via upload
A recommender system, or a recommendation system, is a subclass of information filtering system that seeks to predict the "rating" or "preference" a user would give to an item. They are primarily used in commercial applications.
One file is for best Book  based on author and rating  and another one for best Movie based on ratings and reviews.